c/item_(list|bag): starting to port test to new API

+++ include/pkmn-c/types/item_slot.h
@@ -15,7 +15,7 @@
 typedef struct
 {
     char* item;
-    int amount;
+    size_t amount;
 } pkmn_item_slot_t;
 
 typedef struct
